models/HarmoniDiff/requirements.txt
diffusers==0.38.0
numpy==2.4.4
packaging==26.2
Pillow==12.2.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
transformers==5.8.1
